[dagster-aws] split Pipes tests to separate modules
